Add theoretical Dokka support in Gradle build for test module
This is for the purpose of documenting the Kotlin code that will soon
populate the test module.

Note that as of the time of this commit (12/8/2017 2:10am PST), Gradle
plus Dokka plus Java 9 fails spectacularly, so this doesn't even work.
I'm just putting it in in the hopes that it will be fixed in the not too
distant future.