FIX BUG: Off-by-one error when extracting AttributeType
This is because the enumeration ordinals all run from zero (of course)
but zero in the constant pool metadata bits array means nothing there.
So constant pool tags are -1, -2, ..., -m; and attribute type tags are
1, 2, 3, ..., n; and you have to either offset the map by one (as done
in this commit) or subtract one from the metabits in order to make it
all line up.