FIX BUG: Class parser Code attribute length mismatch
The bug was caused by the fact that the `attributes(int)` method
doesn't tally up its own size anywhere, so when the Code attribute was
asserting that its actual length matched the stated length, it was
forgetting about the size of its sub-attributes.

In this case the solution was to just check the stream position
counter's relative change rather than bothering about passing around
an `int[] actualAttributeLength` array which would have required a lot
of code change. *I wonder if this approach can be used to remove all the
other such `actualAttributeLength` arrays?*

Also remove the FIXME for this problem that I added in commit
**71b54d84**.

Also in this commit: Extend the class parser smoke test with a case that
has a tiny class with a single <init> method.

See: https://github.com/vcschapp/tarro/commit/71b54d8407f8bd1b65b46e3c45