Add JRE 9 version 9.0.1 with fake rt.jar
The `rt.jar` is an "imposter" that I constructed by using the `jimage`
utility to extract the *java.base* module from Java 9's proprietary
`modules` über-file. See the README for details.

I'm not clear what crazy version numbering scheme they're adopting for
Java 9 - is it really 1.9.0.1 or just 9.0.1? Presumably it's the first
one since that would be consistent with the 8 prior major versions and
there's no great reason to switch it up now when they're already
planning on re-switching it to a different version number with the next
release anyway!

Anyways - adopting 1.9.0.1 as the official name for it, and if that
turns out to be wrong, we can change it later.