Move @Represents annotation to be inner class of ConstantPoolTag
The `@Represents` paradigm was unworkable for other packages as
implemented. This is because other packages, such as base.attribute,
contain several enumerations for which an `@Represents` annotation makes
sense. But the annotation concept doesn't support enough genericity to
allow multiple different enumeration type value in the same annotation
type. Hence we need multiple `@Represents` annotations in the same
package. Hence we need to either give them unwieldy names or use some
additional namespacing. We adopt the latter option because the inner
class construct gives us the namespace, and it makes more explicit the
annotation type's association with the enumeration type of which it is
an inner class.