FIX BUG: Wrong bit fiddling in bytecode validator
The old `validateJumps()` code was doing this:

    final long offset = jumpPair & 0xffffffffL;

But obviously the offset can be negative, forexample `ifne -9`. In this
case, assuming the low-order 32 bits of `jumpPair` do in fact represent
-9 (0xfffffff7) then by storing them in the low-order 32 bits of another
`long`, `offset`, you're going to get some giant unsigned value.

Easiest fix is to make sure we're converting back to `int`.

The old `jumpPair()` code was also wrong because for a negative offset,
the OR could result in the putting 0xffffffff into the high-order of
the result, overwriting `sourceInstructionPosition` which then gets
forced to -1.