Refactor base bytecode package to shard opcodes based on operand count
This will make `ByteCodeParser`'s visitors a little bit more natural to
implement because the visitors now receive their opcode as an instance
of the enumeration appropriate to that visitor.

For example instead of:

    interface NoOperandInstructionVisitor {
        void visit(int position, Opcode opcode);
    }

we now have:

    interface NoOperandInstructionVisitor {
        void visit(int position, NoOperandOpcode opcode);
    }

making it easier to see the set of opcodes the visitor needs to
contemplate processing.