Fully document & unit test ATypeValue
This commit adds full Javadoc for `ATypeValue`, plus clean-up, and a
fairly simple unit test covering it to 100%.

It also makes some small changes to `ConstantContainerTest` to make it a
bit more general. Note that `SingleBitMaskContainerTest` is still
relying on the old `visibleFields` property and I haven't yet evaluated
whether it should instead use `uniqueFields`.