Fully document & unit test ConstantPoolTag
This commit adds full Javadoc for `ConstantPoolTag` plus a fairly simple
unit test covering it 100%. It also simplifies renames some of the
enumeration's members to establish a more consistent relationship
between the naming of the enumerators and the naming of the constant
pool tags in the Java Virtual Machine Specification.

Two unit tests in `ConstantPoolTag.kt` fail because of bugs discovered
while implementing the tests. These bugs will be fixed in the two
commits immediately following this one.