Switch to using hamcrest-all to circumvent Java 9 module issues
Before this commit, the project was using two distinct chunks of the
Hamcrest library: hamcrest-core and hamcrest-library. This was no
problem for unit testing the other modules, but for the test module,
which is itself a module, it was as problem.

The problem occurs because in the top-level `build.gradle` file, there's
a hack that puts all the compile-time dependencies on the compiler's
module path because Gradle doesn't currently (or didn't when I started
the project) have a better way of supporting Java 9. Hence in the test
module where the dependencies are compile-time dependencies, both the
Hamcrest libraries are on the module path. But since both hamcrest-core
and hamcrest-library define classes in the `org.hamcrest` package,
we get the following type of error:

    tarro/test/src/main/java/module-info.java:25: error: module io.tarro.test reads package org.hamcrest from both hamcrest.library and hamcrest.core
    module io.tarro.test {
    ^
    12 errors

The solution is just to put one JAR on the classpath, not two, and
that's achieved by using hamcrest-all for the test module. I made the
parallel change to the top-level build.gradle just for symmetry and
simplicity.