Replace TODO with explanatory comment on thread safety
Basically the comment is expressing the fact that the worst case
is in a concurrent scenario before `ALL_OPCODES` is initialized where
either two threads more or less simultaneously read it as `null` and
therefore duplicate the work of initializing the cached value; or where
one thread has completed the initialization entirely and assigned a
value to `ALL_OPCODES` but the write isn't yet visible to another thread
which then duplicates the work.

This rare duplication is probably considerably cheaper than trying to
maintain code to synchronize a fairly trivial static initialization
across threads.