Change class file version of ACC_ANNOTATION flag to Java 5
Previously it was Java 7, which is wrong. Annotations were added in Java
5. If you check out the JVM 5 Specification class file format document
in doc/jvms/JVMS-SE5.0-Ch4-ClassFile.pdf, it's clear that the flag gets
added in Java 5. Moreover, I have verified that the flag existed in Java
5 class files.

In particular, if you pull an `rt.jar` from a Java 5 release, including
the one in this very project's test module, you'll see that the class
`java.lang.Retention` has access flags 0x2601, *aka*

    ACC_PUBLIC | ACC_INTERFACE | ACC_ABSTRACT | ACC_ANNOTATION