Fully document & unit test FlagMixRule, rename related exception
This commit adds full Javadoc for `FlagMixRule` plus a unit test
covering it 100%. It also renames `BadAccessFlagMixException` to the
slightly simpler and more generic name `BadFlagMixException`.

Added Google Guava as a test dependency in the main `build.gradle` file
because the handy set calculation methods are helpful in the unit tests
for `FlagMixRule`.