Delete the base Visitor interface
This interface seems to me to be superfluous.

It's only real value was to add `before()` and `after()` methods that
the `ByteCodeParser` and `ClassParser` classes could call at the
beginning and end of each individual parse operation.

It seems to me that it is presuming a lot to think that this is
generally-applicable functionality which most people will find useful,
and I don't see the point of bloating the library and slowing down the
"optimal case" parse operation simply because a small number of
consumers might want to reset their visitors in between parse
operations. If they need to, that should be something they can handle
on their own!